requirements.txt
asgiref==3.6.0
Django==4.2.1
django-appconf==1.0.5
django-compressor==4.3.1
django-cors-headers==4.0.0
django-storages==1.13.2
gunicorn==20.1.0
jmespath==1.0.1
lazy-object-proxy==1.9.0
Pillow==11.0.0
psycopg2-binary==2.9.6
python-dotenv==1.0.0
python-dateutil==2.8.2
pytz==2023.3
rcssmin==1.1.1
rjsmin==1.2.1
s3transfer==0.6.1
six==1.16.0
sqlparse==0.4.4
urllib3==1.26.15
azure-core==1.26.3
azure-storage-blob==12.14.1
whitenoise==6.4.0